requirements.txt
flask
flask-login
celery
redis
flask-wtf
flask-sqlalchemy
pip-tools
connexion[swagger-ui]